backend/requirements.txt
fastapi==0.128.2
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
pydantic==2.12.5
pydantic[email]==2.12.5
python-dotenv==1.0.1
python-multipart==0.0.9
psycopg2-binary==2.9.11
gunicorn==23.0.0
